backend-fastAPI/fastApiProject-poc/app/config.py
import motor.motor_asyncio
import os
from dotenv import load_dotenv


load_dotenv()

MONGO_URI = os.getenv("MONGO_URI")
DB_NAME = os.getenv("MONGO_DB_NAME")


client = motor.motor_asyncio.AsyncIOMotorClient(MONGO_URI)
database = client[DB_NAME]


students_collection = database.get_collection("students")
books_collection = database.get_collection("books")
